requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
python-multipart==0.0.20
python-dotenv==1.1.0

# Vector search
faiss-cpu==1.10.0
numpy==1.26.4
scikit-learn==1.6.1
llama-index-core==0.12.33
llama-index-vector-stores-faiss==0.3.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
beautifulsoup4==4.13.4
markdown==3.8
pymupdf==1.25.5  # fitz
python-pptx==1.0.2

# Embedding models
sentence-transformers==4.1.0
requests==2.32.3

# DSPy integration
dspy==2.6.19
openai==1.61.0
litellm==1.63.7
packaging==24.2

# CLI
typer==0.15.2
rich==14.0.0

# UI
streamlit==1.44.1
plotly==6.0.1
pandas==2.2.3

# Security
python-jose==3.4.0
passlib==1.7.4
bcrypt==4.3.0
pyotp==2.9.0
qrcode==8.1
pillow==11.2.1  # Required for qrcode

# Testing (development only)
pytest==8.3.5
pytest-cov==6.1.1
